fix: egghunter for win10 wow64
the current egghunter for win10 wow64 fails to loop through the pages.
this is due to some missing PUSH EBX (0) before
NtAccessCheckAndAuditAlarm.
without them, last error will be set to INVALID_NAME.
tested on QuickZip 4.60 SEH exploit on win10.

after fixing the page loop, we noticed that this was not enough.
the egghunter has to reset EBX every loop with XOR EBX, EBX,
otherwise it will stop at offset 1 of the current page.